+++ run.ps1
@@ -1,4 +1,5 @@
 # .\build_cangateway.ps1
+$env:CGO_CFLAGS = ""
 $env:GOARCH = "amd64"
 $env:CGO_ENABLED = "1" 
 $env:GOGC = "100"